+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Draft 2.x.x
+- [Issue 47](https://github.com/lemberg/draft/issues/47) - Add custom Heading field formatter module
 - [Issue 44](https://github.com/lemberg/draft/issues/44) - Add paragraph Compound - Media Bar
 - [Issue 46](https://github.com/lemberg/draft/issues/46) - Add paragraph Simple - CTA
 - [Issue 55](https://github.com/lemberg/draft/issues/55) - Add paragraphs Link - Default and Link - File